+++ src/utils/misc.h
@@ -138,6 +138,9 @@ QString generateRandomString(int length = 8);
 QString makeFileNameRandom(const QString &fileName,
                            const QString &overrideSuffix = "");
 QString rstrip(const QString &str);
+bool isPreviewUseEditorStyles();
+QString previewFontString();
+QString previewCodeFontString();
 }    // namespace Misc
 }    // namespace Utils